Move figshare_group function into commons module
